requirements.txt
RPi.GPIO==0.7.1
pygame==2.5.2
